Fix loading of locales in certain cases
Change the locales path for loading to an absolute path (based on the
main.py file) in order to allow launching the StreamController from any
path. Currently we can only loads it with workdir being the source root
of the project otherwise locales won’t be found and project will refuse
to start.