feat(devcontainers): add devcontainer support to the project (#393)
* feat(devcontainers): add devcontainer support to the project

Add initial devcontainer support to the project, and some basic linting

* fix: refactor to a dockerfile to make proper use of dbus/usb passthrough etc.

By moving to a dockerfile this provides us with ability to correctly use
the newer versions of adwaita, passthrough of usb devices and wayland,
and run a session dbus easily solving a number of devcontainer issues I
came across on bluefin.

* fix: removing linter/other opinionated settings

Removing the opinionated settings from the devcontainer/IDE env to avoid
undesirable misalignment with existing standard.